requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==3.0.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.9.2
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
flake8==6.1.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.25.2
packaging==25.0
pandas==2.1.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyfcm==1.5.4
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.2
redis==5.0.1
requests==2.32.4
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
threadpoolctl==3.6.0
twilio==8.10.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==12.0
yarl==1.20.1
